+++ WebApplicationProjetoAula5/WebApplicationProjetoAula5/FORNECEDOR.cs
@@ -26,7 +26,7 @@ namespace WebApplicationProjetoAula5
         public string telefone { get; set; }
         public string cidade { get; set; }
         public string endereco { get; set; }
-        public Nullable<int> cnpj { get; set; }
+        public Nullable<long> cnpj { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<COMPRA> COMPRA { get; set; }
